requirements.txt
pyrogram
tgcrypto
pyromod
python-dotenv
